refactor: isolate third-party blocking executor
